+++ Examples/Load-ISHBootstrapperContext.ps1
@@ -20,11 +20,14 @@ try
     $data = $client.DownloadString($uri.AbsoluteUri)
 
     $variableName="__ISHBootstrapper_Data__"
-    Set-Variable $variableName -Value ($data| ConvertFrom-Json) -Scope Global -Force
+	$json=$data| ConvertFrom-Json
+    $json | Add-Member NoteProperty "FolderPath" -Value (Split-Path -Parent $JSONPath)
+    $json | Add-Member NoteProperty "FileName" -Value (Split-Path -Leaf $JSONPath)
+    Set-Variable $variableName -Value $json -Scope Global -Force
 
-    Write-Host "ISHBootstrapper initialized from $JSONPath in variable $variableName"
+    Write-Host "ISHBootstrapper initialized from $JSONPath in variable $variableName."
 }
-catch
+finally
 {
-    throw
+
 }